+++ src/main/java/org/example/study_group_service/repository/StudyGroupRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudyGroupRepository extends JpaRepository<AdminRequestEntity, Integer> {}
+public interface StudyGroupRepository extends JpaRepository<AdminRequestEntity, Long> {}
