+++ src/main/resources/db/changelog/changeset/0001-create-tables.sql
@@ -1,12 +1,12 @@
 CREATE TABLE t_role
 (
-    id   SERIAL PRIMARY KEY NOT NULL,
+    id   BIGSERIAL PRIMARY KEY NOT NULL,
     role VARCHAR(100)       NOT NULL
 );
 
 CREATE TABLE t_user
 (
-    id                SERIAL PRIMARY KEY      NOT NULL,
+    id                BIGSERIAL PRIMARY KEY      NOT NULL,
     username          VARCHAR(100)            NOT NULL,
     password          VARCHAR(100)            NOT NULL,
     email             VARCHAR(100),
@@ -15,7 +15,7 @@ CREATE TABLE t_user
 
 CREATE TABLE t_user_roles
 (
-    id      SERIAL PRIMARY KEY NOT NULL,
-    role_id INT,
-    user_id INT
+    id      BIGSERIAL PRIMARY KEY NOT NULL,
+    role_id BIGINT,
+    user_id BIGINT
 );
