+++ src/main/resources/db/changelog/changeset/0004-create-admin-request-table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE admin_request
 (
-    id                SERIAL PRIMARY KEY      NOT NULL,
-    user_id           INT                     NOT NULL,
+    id                BIGSERIAL PRIMARY KEY      NOT NULL,
+    user_id           BIGINT                     NOT NULL,
     creation_time     TIMESTAMP DEFAULT NOW() NOT NULL,
-    processed_user_id INT,
+    processed_user_id BIGINT,
     procession_time   TIMESTAMP,
     status            VARCHAR(50)
 );
